prepare for ipynb3 support
